refactor: put SUPPORTED_LANGS in webview interface
