--- src/langs.ts
@@ -1,38 +1,0 @@
-export const SUPPORTED_LANGS = [
-  "bash",
-  "sh",
-  "c",
-  "clojure",
-  "c++",
-  "c#",
-  "dart",
-  "dockerfile",
-  "elixir",
-  "go",
-  "hack",
-  "html",
-  "java",
-  "javascript",
-  "json",
-  "jsonnet",
-  "julia",
-  "kotlin",
-  "lisp",
-  "lua",
-  "ocaml",
-  "php",
-  "python",
-  "r",
-  "ruby",
-  "rust",
-  "scala",
-  "scheme",
-  "solidity",
-  "sol",
-  "swift",
-  "terraform",
-  "typescript",
-  "vue",
-  "xml",
-  "yaml",
-];
